refactor(docker): rename registry helper file
